+++ src/context.ts
@@ -1,0 +1,5 @@
+import * as vscode from 'vscode';
+
+export const contextContainer: { current: vscode.ExtensionContext | null;  } = {
+    current: null
+};