+++ src/util/imagecache.ts
@@ -1,10 +1,11 @@
 import * as vscode from 'vscode';
 import * as path from 'path';
 import * as fs from 'fs';
-import { parseDds } from './ddsparser';
+import { parseDds, DDS } from './ddsparser';
 import { PNG } from 'pngjs';
-import { parseHoi4File } from './hoiparser';
-import { forEachNodeValue, getPropertyNodes, getStringPropertyOrUndefined } from './hoinodeutils';
+import { parseHoi4File } from '../hoiformat/hoiparser';
+import { forEachNodeValue, getPropertyNodes, getStringPropertyOrUndefined } from '../hoiformat/hoinodeutils';
+import { getSpriteTypes } from '../hoiformat/spritetype';
 
 interface Icon {
     uri: string;
@@ -25,7 +26,7 @@ export function getFocusIcon(name: string): Icon | null {
 
     const ddsPath = gfxMap[name] ?? 'gfx/interface/goals/goal_unknown.dds';
 
-    const conf = vscode.workspace.getConfiguration('hoi4focustreepreview');
+    const conf = vscode.workspace.getConfiguration('hoi4modutilities');
     const installPath = conf.installPath;
 
     if (!installPath) {
@@ -40,34 +41,13 @@ export function getFocusIcon(name: string): Icon | null {
     try {
         const buffer = fs.readFileSync(fullpath).buffer; // (await new Promise<Buffer>((resolve, reject) => fs.readFile(fullpath, (e, d) => e ? reject(e) : resolve(d)))).buffer;
         const dds = parseDds(buffer);
-
-        const img = dds.mainSurface;
-
-        const png = new PNG({ width: img.width, height: img.height });
-        const imgbuffer = img.pixelSizeInByte === 1 ? new Uint8Array(buffer, img.offset, img.length) : (
-            img.pixelSizeInByte === 2 ? new Uint16Array(buffer, img.offset, img.length / 2) :
-            new Uint32Array(buffer, img.offset, img.length / 4)
-        );
-
-        const masks = [dds.header.ddspf.dwRBitMask, dds.header.ddspf.dwGBitMask, dds.header.ddspf.dwBBitMask, dds.header.ddspf.dwABitMask];
-        const moves = masks.map(end0count);
-        const scales = masks.map(v => 255 / ((1 << all1Count(v)) - 1));
-        
-        for (let y = 0; y < png.height; y++) {
-            for (let x = 0; x < png.width; x++) {
-                let idx = (png.width * y + x) << 2;
-            
-                for (let i = 0; i < 4; i++) {
-                    png.data[idx + i] = ((imgbuffer[idx / 4] & masks[i]) >> moves[i]) * scales[i];
-                }
-            }
-        }
+        const png = ddsToPng(dds);
 
         const pngBuffer = PNG.sync.write(png);
         const result: Icon = {
             uri: 'data:image/png;base64,' + pngBuffer.toString('base64'),
-            width: img.width,
-            height: img.height
+            width: png.width,
+            height: png.height
         };
 
         cache[name] = result;
@@ -98,12 +78,42 @@ function all1Count(v: number): number {
 	return v;
 }
 
+function ddsToPng(dds: DDS): PNG {
+    const img = dds.mainSurface;
+    const buffer = dds.buffer;
+
+    const png = new PNG({ width: img.width, height: img.height });
+    const imgbuffer = img.pixelSizeInByte === 1 ? new Uint8Array(buffer, img.offset, img.length) : (
+        img.pixelSizeInByte === 2 ? new Uint16Array(buffer, img.offset, img.length / 2) :
+        new Uint32Array(buffer, img.offset, img.length / 4)
+    );
+
+    const masks = [dds.header.ddspf.dwRBitMask, dds.header.ddspf.dwGBitMask, dds.header.ddspf.dwBBitMask, dds.header.ddspf.dwABitMask];
+    const moves = masks.map(end0count);
+    const scales = masks.map(v => 255 / ((1 << all1Count(v)) - 1));
+    
+    for (let y = 0; y < png.height; y++) {
+        for (let x = 0; x < png.width; x++) {
+            let idx = (png.width * y + x) << 2;
+        
+            for (let i = 0; i < 4; i++) {
+                png.data[idx + i] = ((imgbuffer[idx / 4] & masks[i]) >> moves[i]) * scales[i];
+                if (i === 3 && (dds.header.ddspf.dwFlags & 1) === 0) {
+                    png.data[idx + i] = 255;
+                }
+            }
+        }
+    }
+
+    return png;
+}
+
 function loadGfxMap(): void {
     if (gfxMapLoaded) {
         return;
     }
 
-    const conf = vscode.workspace.getConfiguration('hoi4focustreepreview');
+    const conf = vscode.workspace.getConfiguration('hoi4modutilities');
     const installPath = conf.installPath;
 
     if (!installPath) {
@@ -119,28 +129,13 @@ function loadGfxMap(): void {
         const buffer = fs.readFileSync(fullpath); // (await new Promise<Buffer>((resolve, reject) => fs.readFile(fullpath, (e, d) => e ? reject(e) : resolve(d)))).buffer;
         const gfx = buffer.toString('utf-8');
         const node = parseHoi4File(gfx);
+        const spriteTypes = getSpriteTypes(node);
 
-        forEachNodeValue(node, stsNode => {
-            if (stsNode.name !== 'spriteTypes') {
-                return;
-            }
-
-            const sprites = getPropertyNodes(stsNode, 'SpriteType');
-            for (const sprite of sprites) {
-                const name = getStringPropertyOrUndefined(sprite, 'name');
-                const texturefile = getStringPropertyOrUndefined(sprite, 'texturefile');
-                if (name && texturefile) {
-                    gfxMap[name] = texturefile;
-                }
-            }
-        });
+        spriteTypes.forEach(st => gfxMap[st.name] = st.texturefile);
 
     } catch (e) {
         console.error(e);
     }
-
-    'interface/goals.gfx';
-
     
     gfxMapLoaded = true;
 }