+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 # Change Log
 
-All notable changes to the "hoi4focustreepreview" extension will be documented in this file.
+All notable changes to the "hoi4modutilities" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
 
-- Initial release
+- Preview focus tree
+    - Add preview page for focus tree file (in `common/national_focus`).
+    - Display focus icon at their position and show lines for dependencies.
+    - Click focus at preview page will navigate to the focus tag in document.
+    - Hover focus will show focus ID and absolute position.
+    - Focus with `allow_branch` will have a option to hide it.
+    - Preview view will be updated when focus file updated.
+    - Can preview `shared_focus`.